refactor reviews panel into component
